+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added api endpoint to for targeted custom field filling #17
 - added a way to automatically configure workflows for all supported custom fields, need to be enabled by user explicitly #20
 - updated token sampling and context management to be grammar aware and only let the model generate non deterministic parts of the output
+- updated benchmarking to track token generation performance statistics
+- add script to evalute token generation stats and compare models
+
+### Fixed
+- fixed bug that prevented benchmark continuation for unfinished models
 
 ## [ 0.4.0 - 2026-06-19]